fix & refactor: /analyzechests gui no work
Fixed `/analyzechests gui` ran as a command not working. Refactor it to be better performance along side the fix.

BREAKING CHANGE: There are no known breaking changes however there is potential if say lag, spam to break it or cause it to open multiple gui back to back thereby needing to exit gui multiple times.
  I tested minimally and there were no errors.

Signed-off-by: Kami <harshimymommy@gmail.com>